# hq/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Zako3 HQ API",
    "description": "RESTful API for Zako3 platform - manage users, communities (taps), and bot settings",
    "version": "1.0.0",
    "contact": {
      "name": "Zako3 Team"
    },
    "license": {
      "name": "ISC"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from authentication provider"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "summary": "Get current user profile",
        "tags": [
          "users"
        ],
        "description": "Get the authenticated user's full profile including private information",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "Discord user ID"
                        },
                        "username": {
                          "type": "string",
                          "description": "Username"
                        },
                        "displayName": {
                          "description": "Display name",
                          "type": "string"
                        },
                        "avatarUrl": {
                          "description": "Avatar URL",
                          "type": "string"
                        },
                        "discriminator": {
                          "description": "User discriminator",
                          "type": "string"
                        },
                        "isBanned": {
                          "type": "boolean",
                          "description": "Whether user is banned"
                        },
                        "createdAt": {
                          "type": "string",
                          "description": "Account creation timestamp"
                        },
                        "lastSeenAt": {
                          "description": "Last seen timestamp",
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "username",
                        "isBanned",
                        "createdAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Error code identifier"
                        },
                        "message": {
                          "type": "string",
                          "description": "Human-readable error message"
                        },
                        "details": {
                          "description": "Additional error details"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Error code identifier"
                        },
                        "message": {
                          "type": "string",
                          "description": "Human-readable error message"
                        },
                        "details": {
                          "description": "Additional error details"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Error code identifier"
                        },
                        "message": {
                          "type": "string",
                          "description": "Human-readable error message"
                        },
                        "details": {
                          "description": "Additional error details"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{userId}": {
      "get": {
        "summary": "Get user public profile",
        "tags": [
          "users"
        ],
        "description": "Get a public user profile by user ID",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "userId",
            "required": true,
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "Discord user ID"
                        },
                        "username": {
                          "type": "string",
                          "description": "Username"
                        },
                        "displayName": {
                          "description": "Display name",
                          "type": "string"
                        },
                        "avatarUrl": {
                          "description": "Avatar URL",
                          "type": "string"
                        },
                        "discriminator": {
                          "description": "User discriminator",
                          "type": "string"
                        },
                        "isBanned": {
                          "type": "boolean",
                          "description": "Whether user is banned"
                        },
                        "createdAt": {
                          "type": "string",
                          "description": "Account creation timestamp"
                        },
                        "lastSeenAt": {
                          "description": "Last seen timestamp",
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "username",
                        "isBanned",
                        "createdAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Error code identifier"
                        },
                        "message": {
                          "type": "string",
                          "description": "Human-readable error message"
                        },
                        "details": {
                          "description": "Additional error details"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Error code identifier"
                        },
                        "message": {
                          "type": "string",
                          "description": "Human-readable error message"
                        },
                        "details": {
                          "description": "Additional error details"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://api.zako3.com",
      "description": "Production server"
    }
  ],
...
}